Staging: vt6655: remove custom ULONGLONG typedef

+++ drivers/staging/vt6655/bssdb.h
@@ -212,7 +212,7 @@ typedef struct tagKnownNodeDB {
     WORD            wEnQueueCnt;
 
     BOOL            bOnFly;
-    ULONGLONG       KeyRSC;
+    unsigned long long       KeyRSC;
     BYTE            byKeyIndex;
     DWORD           dwKeyIndex;
     BYTE            byCipherSuite;
